feat!: allow to select the number of attempts
BREAKING CHANGES: The GallowsArtist constructor now requires no arguments. The NumberOfAttempts value is passed to GallowsArtist using the getCurrGallowsState method argument. These changes are necessary because the number of attempts is determined by the user now and is not defined in the code any longer;
Allow user to enter the number of attempts(in other words, maximum number of mistakes).